+++ src/ldc/filter/__init__.py
@@ -22,7 +22,7 @@ from ._replace_patterns import ReplacePatterns
 from ._reset_ids import ResetIDs
 from ._skip_duplicate_ids import SkipDuplicateIDs
 from ._skip_duplicate_text import SkipDuplicateText
-from ._split import Split
+from ._split_records import SplitRecords
 from ._tee import Tee
 from ._text_length import TextLength
 from ._text_stats import TextStatistics