+++ src/ldc/filter/_split_records.py
@@ -12,7 +12,7 @@ META_SPLIT = "split"
 """ the key for storing the split name in the meta-data. """
 
 
-class Split(Filter):
+class SplitRecords(Filter):
     """
     Splits the incoming records into the specified split ratios by setting the 'split' meta-data value. Also stores the split names in the current session.
     """
@@ -43,7 +43,7 @@ class Split(Filter):
         :return: the name
         :rtype: str
         """
-        return "split"
+        return "split-records"
 
     def description(self) -> str:
         """